serving_model_code/requirements.txt
flask
gunicorn
pandas
xgboost
scikit-learn
google-cloud-storage